Moved FileStorageTable methods into FileStorageBehavior.
